Apply language to context before creating Notifications
