+++ k9mail/src/main/java/com/fsck/k9/K9UICommon.java
@@ -1,0 +1,29 @@
+package com.fsck.k9;
+
+
+import java.util.Locale;
+
+import android.content.Context;
+import android.content.res.Configuration;
+import android.content.res.Resources;
+import android.text.TextUtils;
+
+
+public class K9UICommon {
+    public static void setLanguage(Context context, String language) {
+        Locale locale;
+        if (TextUtils.isEmpty(language)) {
+            locale = Locale.getDefault();
+        } else if (language.length() == 5 && language.charAt(2) == '_') {
+            // language is in the form: en_US
+            locale = new Locale(language.substring(0, 2), language.substring(3));
+        } else {
+            locale = new Locale(language);
+        }
+
+        Configuration config = new Configuration();
+        config.locale = locale;
+        Resources resources = context.getResources();
+        resources.updateConfiguration(config, resources.getDisplayMetrics());
+    }
+}
