+++ k9mail/src/main/java/com/fsck/k9/activity/K9PreferenceActivity.java
@@ -6,12 +6,13 @@ import android.preference.Preference;
 import android.preference.PreferenceActivity;
 
 import com.fsck.k9.K9;
+import com.fsck.k9.K9UICommon;
 
 
 public abstract class K9PreferenceActivity extends PreferenceActivity {
     @Override
     public void onCreate(Bundle icicle) {
-        K9ActivityCommon.setLanguage(this, K9.getK9Language());
+        K9UICommon.setLanguage(this, K9.getK9Language());
         setTheme(K9.getK9ThemeResourceId());
         super.onCreate(icicle);
     }
