+++ k9mail/src/test/java/com/fsck/k9/mailstore/MessageViewInfoExtractorTest.java
@@ -12,6 +12,7 @@ import android.app.Application;
 
 import com.fsck.k9.GlobalsHelper;
 import com.fsck.k9.K9RobolectricTestRunner;
+import com.fsck.k9.K9UICommon;
 import com.fsck.k9.activity.K9ActivityCommon;
 import com.fsck.k9.message.html.HtmlSanitizer;
 import com.fsck.k9.message.html.HtmlSanitizerHelper;
@@ -217,7 +218,7 @@ public class MessageViewInfoExtractorTest {
 
     @Test
     public void testTextPlusRfc822Message() throws MessagingException {
-        K9ActivityCommon.setLanguage(context, "en");
+        K9UICommon.setLanguage(context, "en");
         Locale.setDefault(Locale.US);
         TimeZone.setDefault(TimeZone.getTimeZone("GMT+01:00"));
